+++ routes/web.php
@@ -2,8 +2,14 @@
 
 use App\Http\Controllers\AutorController;
 use App\Http\Controllers\DashboardController;
+use App\Http\Controllers\DevolucionController;
+use App\Http\Controllers\HistorialController;
 use App\Http\Controllers\LibrosController;
-
+use App\Http\Controllers\PrestamoController;
+use App\Http\Controllers\RolController;
+use App\Http\Controllers\RolesController;
+use App\Http\Controllers\UsuarioController;
+use GuzzleHttp\Middleware;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Facades\Route;
@@ -15,25 +21,46 @@ Auth::routes();
 
 Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard')->middleware('auth');
 
-Route::get('/libros', [LibrosController::class, 'index'])->name('libro.index')->middleware('auth');
-Route::post('/libros', [LibrosController::class, 'index'])->name('libro.index')->middleware('auth');
-Route::get('/libros/nuevo', [LibrosController::class, 'registerStore'])->name('libro.new')->middleware('auth');
-Route::post('/libros/nuevo', [LibrosController::class, 'register'])->name('libro.new')->middleware('auth');
-Route::get('/libros/edit/{libro}', [LibrosController::class, 'edit'])->name('libro.edit')->middleware('auth');
-Route::put('/libros/{libro}', [LibrosController::class, 'update'])->name('libro.update');
-Route::delete('/libros/delete/{id}', [LibrosController::class, 'delete'])->name('libro.delete');
+Route::get('/libros', [LibrosController::class, 'index'])->name('libro.index')->middleware('can:isAdminOrTrabajador');
+Route::post('/libros', [LibrosController::class, 'index'])->name('libro.index')->middleware('can:isAdminOrTrabajador');
+Route::get('/libros/nuevo', [LibrosController::class, 'registerStore'])->name('libro.new')->middleware('can:isAdminOrTrabajador');
+Route::post('/libros/nuevo', [LibrosController::class, 'register'])->name('libro.new')->middleware('can:isAdminOrTrabajador');
+Route::get('/libros/edit/{libro}', [LibrosController::class, 'edit'])->name('libro.edit')->middleware('can:isAdminOrTrabajador');
+Route::put('/libros/{libro}', [LibrosController::class, 'update'])->name('libro.update')->middleware('can:isAdminOrTrabajador');
+Route::delete('/libros/delete/{id}', [LibrosController::class, 'delete'])->name('libro.delete')->middleware('can:isAdminOrTrabajador');
+Route::post('/libros/plantilla', [LibrosController::class, 'addToPlantilla'])->name('libro.plantilla');
 
-Route::get('/autor', [AutorController::class, 'index'])->name('autor.index')->middleware('auth');
-Route::post('/autor', [AutorController::class, 'index'])->name('autor.index')->middleware('auth');
-Route::put('/autor', [AutorController::class, 'update'])->name('autor.update');
-Route::delete('/autor/delete/{id}', [AutorController::class, 'delete'])->name('autor.delete');
+Route::post('/libros/baul/{user}/{book}',[LibrosController::class, 'addToBaul'])->name('libro.addToBaul');
+Route::post('/libros/dbaul/{user}/{book}',[LibrosController::class, 'deleteFromBaul'])->name('libro.deleteFromBaul');
 
-Route::view('/prestamos', 'trabajador.prestamo.index')->name('prestamo.index')->middleware('auth');
+Route::get('/autor', [AutorController::class, 'index'])->name('autor.index')->middleware('can:isAdminOrTrabajador');
+Route::post('/autor', [AutorController::class, 'index'])->name('autor.index')->middleware('can:isAdminOrTrabajador');
+Route::put('/autor', [AutorController::class, 'update'])->name('autor.update')->middleware('can:isAdminOrTrabajador');
+Route::delete('/autor/delete/{id}', [AutorController::class, 'delete'])->name('autor.delete')->middleware('can:isAdminOrTrabajador');
 
-Route::view('/devoluciones', 'trabajador.devolucion.index')->name('devolucion.index')->middleware('auth');
+Route::get('/usuarios', [UsuarioController::class, 'index'])->name('usuario.index')->middleware('can:isAdmin');
+Route::post('/usuarios', [UsuarioController::class, 'index'])->name('usuario.index')->middleware('can:isAdmin');
+Route::get('/usuarios/crear', [UsuarioController::class, 'create'])->name('usuario.create')->middleware('can:isAdmin');
+Route::post('/usuarios/guardar',[UsuarioController::class, 'store'])->name('usuario.save')->middleware('can:isAdmin');
+Route::get('/usuarios/roles', [RolesController::class, 'index'])->name('rol.index')->middleware('can:isAdmin');
+Route::get('/usuarios/roles/crear', [RolesController::class, 'create'])->name('rol.create')->middleware('can:isAdmin');
+Route::post('/usuarios/roles/new', [RolesController::class, 'store'])->name('rol.store')->middleware('can:isAdmin');
+Route::get('/usuarios/roles/{rol}', [RolesController::class, 'show'])->name('rol.show')->middleware('can:isAdmin');
+Route::get('/usuarios/roles/{rol}/edit', [RolesController::class, 'edit'])->name('rol.edit')->middleware('can:isAdmin');
+Route::put('/usuarios/roles/{rol}', [RolesController::class, 'update'])->name('rol.update')->middleware('can:isAdmin');
+Route::delete('/usuarios/roles/delete', [RolesController::class, 'destroy'])->name('rol.delete')->middleware('can:isAdmin');
+Route::get('/usuarios/{user}', [UsuarioController::class, 'edit'])->name('usuario.edit')->middleware('can:isAdmin');
+Route::put('/usuarios/update/{user}', [UsuarioController::class, 'update'])->name('usuario.update')->middleware('can:isAdmin');
+Route::delete('/usuarios/delete/{user}', [UsuarioController::class, 'destroy'])->name('usuario.delete')->middleware('can:isAdmin');
 
-Route::view('/trabajador', 'trabajador.trabajador.index')->name('trabajador.index')->middleware('auth');
+Route::get('/prestamos', [PrestamoController::class, 'index'])->name('prestamo.index')->middleware('can:isAdminOrTrabajadorOrProfesor');
+Route::post('/prestamos', [PrestamoController::class, 'index'])->name('prestamo.index')->middleware('can:isAdminOrTrabajadorOrProfesor');
+Route::get('/prestamos/baul',[PrestamoController::class, 'viewBaul'])->name('prestamo.baul');
+Route::post('/prestamos/baul/finalizar', [PrestamoController::class, 'store'])->name('prestamo.store');
+Route::post('/prestamos/dbaul/{user}/{book}',[PrestamoController::class,'deleteFromBaul'])->name('prestamo.deleteFromBaul');
+Route::get('/prestamos/plantillas',[PrestamoController::class, 'viewListPlantillas'])->name('plantilla.index');
+Route::post('/prestamos/plantillas/save', [PrestamoController::class, 'savePlantilla'])->name('plantilla.save');
 
-Route::view('/estudiante', 'trabajador.estudiante.index')->name('estudiante.index')->middleware('auth');
+Route::get('/devoluciones', [DevolucionController::class, 'index'])->name('devolucion.index')->middleware('can:isAdminOrTrabajadorOrProfesor');
 
-Route::view('/historial', 'trabajador.historial.index')->name('historial.index')->middleware('auth');
+Route::get('/historial', [HistorialController::class, 'index'])->name('historial.index')->middleware('can:isAdminOrTrabajador');
